Fix specs and migrations

+++ db/migrate/20140403013413_remove_wished_products_for_deleted_variants.rb
@@ -1,7 +1,9 @@
 class RemoveWishedProductsForDeletedVariants < ActiveRecord::Migration
   def up
-    Spree::WishedProduct.includes(:variant).find_each do |wished_product|
-      wished_product.destroy unless wished_product.variant
+    if Object.const_defined?('Spree::WishedProduct')
+      Spree::WishedProduct.includes(:variant).find_each do |wished_product|
+        wished_product.destroy unless wished_product.variant
+      end
     end
   end
